Transformed into a project with web services (final)
