--- src/BL/BLTransactionDetails.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Data;
-using System.Configuration;
-using System.Web;
-using System.Xml.Linq;
-using Model;
-using DL;
-
-namespace BL
-{
-    class BLTransactionDetails
-    {
-    }
-}
