+++ src/MenteeAssignment.sln
@@ -7,11 +7,11 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MenteeAssignment", "MenteeA
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Model", "Model\Model.csproj", "{3DEB6090-D8FE-495B-9D7F-AD1C7F3C00AE}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BL", "BL\BL.csproj", "{A1AC58A1-FAC7-402E-AD9B-FC9B90B620F8}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DL", "DL\DL.csproj", "{9E808D7A-C96C-4E72-8D22-697174E77AF3}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DBAccess", "DBAccess\DBAccess.csproj", "{2A309AD2-7F58-422F-9D1F-92CE85DA22B0}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WCF", "WCF\WCF.csproj", "{E748D097-B17D-4C11-BDE1-CBAC22C3F7B8}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,10 +27,6 @@ Global
 		{3DEB6090-D8FE-495B-9D7F-AD1C7F3C00AE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3DEB6090-D8FE-495B-9D7F-AD1C7F3C00AE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3DEB6090-D8FE-495B-9D7F-AD1C7F3C00AE}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A1AC58A1-FAC7-402E-AD9B-FC9B90B620F8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A1AC58A1-FAC7-402E-AD9B-FC9B90B620F8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A1AC58A1-FAC7-402E-AD9B-FC9B90B620F8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A1AC58A1-FAC7-402E-AD9B-FC9B90B620F8}.Release|Any CPU.Build.0 = Release|Any CPU
 		{9E808D7A-C96C-4E72-8D22-697174E77AF3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{9E808D7A-C96C-4E72-8D22-697174E77AF3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9E808D7A-C96C-4E72-8D22-697174E77AF3}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -39,6 +35,10 @@ Global
 		{2A309AD2-7F58-422F-9D1F-92CE85DA22B0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2A309AD2-7F58-422F-9D1F-92CE85DA22B0}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2A309AD2-7F58-422F-9D1F-92CE85DA22B0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E748D097-B17D-4C11-BDE1-CBAC22C3F7B8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E748D097-B17D-4C11-BDE1-CBAC22C3F7B8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E748D097-B17D-4C11-BDE1-CBAC22C3F7B8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E748D097-B17D-4C11-BDE1-CBAC22C3F7B8}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
