+++ src/MenteeAssignment/UI/Maintenance/Customer.aspx.cs
@@ -11,7 +11,7 @@ using System.Web.Security;
 using System.Web.UI.WebControls.WebParts;
 using System.Web.UI.HtmlControls;
 using System.Collections.Generic;
-using BL;
+using WCF;
 using Model;
 
 namespace MenteeAssignment.UI.Maintenance
@@ -88,8 +88,8 @@ namespace MenteeAssignment.UI.Maintenance
                 if (ViewState["OrderByCustomer"] == null) { ViewState["OrderByCustomer"] = string.Empty; }
                 if (ViewState["DT_Customer"] == null)
                 {
-                    BLCustomer BC = new BLCustomer();
-                    ViewState["DT_Customer"] = BC.getAllCustomer(ViewState["WhereCustomer"].ToString(), ViewState["OrderByCustomer"].ToString());
+                    WCFCustomer WC = new WCFCustomer();
+                    ViewState["DT_Customer"] = WC.getAllCustomer(ViewState["WhereCustomer"].ToString(), ViewState["OrderByCustomer"].ToString());
                 }
                 return (DataTable)ViewState["DT_Customer"];
             }
@@ -113,7 +113,7 @@ namespace MenteeAssignment.UI.Maintenance
             switch (e.CommandName)
             {
                 case "Edit":
-                    FillFieldsCustomer(BLCustomer.getCustomerByID(Convert.ToString(e.CommandArgument)));
+                    FillFieldsCustomer(WCFCustomer.getCustomerByID(Convert.ToString(e.CommandArgument)));
                     ManageFieldsCustomer(true, false);
 
                     imgbtnAddCustomer.Visible = false;
@@ -123,7 +123,7 @@ namespace MenteeAssignment.UI.Maintenance
                     pnlCustomer.Enabled = false;
                     break;
                 case "Delete":
-                    BLCustomer.delete(Convert.ToString(e.CommandArgument));
+                    WCFCustomer.delete(Convert.ToString(e.CommandArgument));
                     imgbtnAddCustomer.Visible = true;
                     imgbtnSaveCustomer.Visible = false;
                     imgbtnCancelCustomer.Visible = false;
@@ -147,17 +147,17 @@ namespace MenteeAssignment.UI.Maintenance
         }
         protected void imgbtnSaveCustomer_Clicked(object sender, EventArgs e)
         {
-            bool UpdateFlag = Convert.ToBoolean(BLCustomer.updateFlag(txtID.Text));
+            bool UpdateFlag = Convert.ToBoolean(WCFCustomer.updateFlag(txtID.Text));
 
             if (UpdateFlag == true)
             {
-                BLCustomer.update(GetFieldValuesCustomer());
+                WCFCustomer.update(GetFieldValuesCustomer());
                 ViewState.Remove("DT_Customer");
                 BindPageDataSourceCustomer("");
             }
             else
             {
-                BLCustomer.add(GetFieldValuesCustomer());
+                WCFCustomer.add(GetFieldValuesCustomer());
                 ViewState.Remove("DT_Customer");
                 BindPageDataSourceCustomer("");
             }
@@ -378,7 +378,7 @@ namespace MenteeAssignment.UI.Maintenance
             ViewState.Remove("WhereCustomer");
 
             if (txtSearchCustomer.Text.Trim().Length > 0)
-                ViewState["WhereCustomer"] = BLCustomer.determineHeader(lblSearchCustomer.Text, txtSearchCustomer.Text);
+                ViewState["WhereCustomer"] = WCFCustomer.determineHeader(lblSearchCustomer.Text, txtSearchCustomer.Text);
 
             ViewState.Remove("DT_Customer");
             BindPageDataSourceCustomer("");
@@ -393,7 +393,7 @@ namespace MenteeAssignment.UI.Maintenance
             lblSearchCustomer.Text = e.CommandArgument.ToString();
 
             ViewState["OrderByCustomer"] = "";
-            ViewState["OrderByCustomer"] = BLCustomer.determineSort(lblSearchCustomer.Text, LastSortDirection());
+            ViewState["OrderByCustomer"] = WCFCustomer.determineSort(lblSearchCustomer.Text, LastSortDirection());
 
             BindPageDataSourceCustomer("");
 
@@ -413,8 +413,8 @@ namespace MenteeAssignment.UI.Maintenance
                 if (ViewState["OrderByTransaction"] == null) { ViewState["OrderByTransaction"] = string.Empty; }
                 if (ViewState["DT_Transaction"] == null)
                 {
-                    BLTransactions BT = new BLTransactions();
-                    ViewState["DT_Transaction"] = BT.getAllTransaction(ViewState["WhereTransaction"].ToString(), ViewState["OrderByTransaction"].ToString());
+                    WCFTransactions WT = new WCFTransactions();
+                    ViewState["DT_Transaction"] = WT.getAllTransaction(ViewState["WhereTransaction"].ToString(), ViewState["OrderByTransaction"].ToString());
                 }
                 return (DataTable)ViewState["DT_Transaction"];
             }
@@ -438,7 +438,7 @@ namespace MenteeAssignment.UI.Maintenance
             switch (e.CommandName)
             {
                 case "Edit":
-                    FillFieldsTransaction(BLTransactions.getTransactionByID(Convert.ToString(e.CommandArgument)));
+                    FillFieldsTransaction(WCFTransactions.getTransactionByID(Convert.ToString(e.CommandArgument)));
                     ManageFieldsTransaction(true, false);
 
                     imgbtnAddTransaction.Visible = false;
@@ -448,7 +448,7 @@ namespace MenteeAssignment.UI.Maintenance
                     pnlTransaction.Enabled = false;
                     break;
                 case "Delete":
-                    BLTransactions.delete(Convert.ToString(e.CommandArgument));
+                    WCFTransactions.delete(Convert.ToString(e.CommandArgument));
                     imgbtnAddTransaction.Visible = true;
                     imgbtnSaveTransaction.Visible = false;
                     imgbtnCancelTransaction.Visible = false;
@@ -472,17 +472,17 @@ namespace MenteeAssignment.UI.Maintenance
         }
         protected void imgbtnSaveTransaction_Clicked(object sender, EventArgs e)
         {
-            bool UpdateFlag = Convert.ToBoolean(BLTransactions.updateFlag(txtTransactionID.Text));
+            bool UpdateFlag = Convert.ToBoolean(WCFTransactions.updateFlag(txtTransactionID.Text));
 
             if (UpdateFlag == true)
             {
-                BLTransactions.update(GetFieldValuesTransaction());
+                WCFTransactions.update(GetFieldValuesTransaction());
                 ViewState.Remove("DT_Transaction");
                 BindPageDataSourceTransaction("");
             }
             else
             {
-                BLTransactions.add(GetFieldValuesTransaction());
+                WCFTransactions.add(GetFieldValuesTransaction());
                 ViewState.Remove("DT_Transaction");
                 BindPageDataSourceTransaction("");
             }
@@ -703,7 +703,7 @@ namespace MenteeAssignment.UI.Maintenance
             ViewState.Remove("WhereTransaction");
 
             if (txtSearchTransaction.Text.Trim().Length > 0)
-                ViewState["WhereTransaction"] = BLTransactions.determineHeader(lblSearchTransaction.Text, txtSearchTransaction.Text);
+                ViewState["WhereTransaction"] = WCFTransactions.determineHeader(lblSearchTransaction.Text, txtSearchTransaction.Text);
 
             ViewState.Remove("DT_Transaction");
             BindPageDataSourceTransaction("");
@@ -718,7 +718,7 @@ namespace MenteeAssignment.UI.Maintenance
             lblSearchTransaction.Text = e.CommandArgument.ToString();
 
             ViewState["OrderByTransaction"] = "";
-            ViewState["OrderByTransaction"] = BLCustomer.determineSort(lblSearchTransaction.Text, LastSortDirection());
+            ViewState["OrderByTransaction"] = WCFTransactions.determineSort(lblSearchTransaction.Text, LastSortDirection());
 
             BindPageDataSourceTransaction("");
