+++ src/WCF/WCFCustomer.svc
@@ -1,0 +1,1 @@
+﻿<%@ ServiceHost Language="C#" Debug="true" Service="WCF.WCFCustomer" CodeBehind="WCFCustomer.svc.cs" %>
