+++ src/WCF/WCFTransactions.svc
@@ -1,0 +1,1 @@
+﻿<%@ ServiceHost Language="C#" Debug="true" Service="WCF.WCFTransactions" CodeBehind="WCFTransactions.svc.cs" %>
